requirements.txt
Django==2.1.7
djangorestframework==3.9.1
drf-yasg>=1.17.1,<1.18
psycopg2>=2.8.4,<2.9.0
django-cors-headers
django-model-utils>=4.0.0,<4.1.0
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==19.9.0
pytz==2018.9
whitenoise>=4.1.2
django-filter>2.3.0


# aioredis==1.3.0
# asgiref==3.2.2
# asn1crypto==0.24.0
# async-timeout==3.0.1
# attrs==19.1.0
# autobahn==19.9.3
# Automat==0.7.0
# certifi==2019.6.16
# cffi==1.12.3
# channels==2.3.0
# channels-redis==2.4.0
# chardet==3.0.4
# configparser==3.7.4
# constantly==15.1.0
# cryptography==2.7
# daphne==2.3.0
# defusedxml==0.6.0
# Django==2.2.4
# django-cors-headers==3.1.0
# djangorestframework==3.10.2
# djangorestframework-jwt==1.11.0
# djangorestframework-simplejwt==4.3.0
# hiredis==1.0.0
# hyperlink==19.0.0
# idna==2.8
# incremental==17.5.0
# msgpack==0.6.2
# numpy==1.17.0
# oauthlib==3.0.2
# pandas==0.25.0
# Pillow==6.2.0
# # pkg-resources==0.0.0
# psycopg2==2.8.3
# pycparser==2.19
# PyHamcrest==1.9.0
# PyJWT==1.7.1
# python-dateutil==2.8.0
# python3-openid==3.1.0
# pytz==2019.2
# requests==2.22.0
# requests-oauthlib==1.2.0
# six==1.12.0
# social-auth-app-django==3.1.0
# social-auth-core==3.2.0
# sqlparse==0.3.0
# Twisted==19.7.0
# txaio==18.8.1
# urllib3==1.25.3
# uuid==1.30
# zope.interface==4.6.0

